Fix Tests failing on GitHub Actions - Increase Jest default timeout to 15000 ms (#36)
After GitHub Actions updates the runner (v2.315.0), it seems the tests take much more time to run. I have been monitoring and running the pull request checks workflow in parallel on Circle CI and almost have no timeout issue. It's maybe one in 10. So I decided to set the `testTimeout` to 15000 and ran the tests a few times. Now, it seems to be consistently passing.